+++ src/__tests__/XHubSignatureMiddleware.test.js
@@ -1,5 +1,5 @@
-const { WebhookMiddleware } = require('../middlewares')
-const webhookMiddleware = new WebhookMiddleware()
+const { XHubSignatureMiddleware } = require('../middlewares')
+const xHubSignatureMiddleware = new XHubSignatureMiddleware()
 const { invalidParamError, missingParamError, missingBodyContent } = require('../helpers/Errors')
 const Mocks = require('./helpers/Mocks')
 const mocks = new Mocks()
@@ -14,7 +14,7 @@ describe("WEBHOOK Middlware Tests", () => {
       const req = mocks.mockReq(leadgen, null, null, null, {})
       const next = mocks.mockNext()
       const { error } = missingParamError('x-hub-signature')
-      await webhookMiddleware.checkSignature(req, res, next)
+      await xHubSignatureMiddleware.checkSignature(req, res, next)
       expect(res.status).toHaveBeenCalledWith(400)
       expect(res.json).toHaveBeenCalledWith(error)
     })
@@ -26,7 +26,7 @@ describe("WEBHOOK Middlware Tests", () => {
       const req = mocks.mockReq(null, null, null, null, header)
       const next = mocks.mockNext()
       const { error } = missingBodyContent()
-      await webhookMiddleware.checkSignature(req, res, next)
+      await xHubSignatureMiddleware.checkSignature(req, res, next)
       expect(res.status).toHaveBeenCalledWith(400)
       expect(res.json).toHaveBeenCalledWith(error)
     })
@@ -39,7 +39,7 @@ describe("WEBHOOK Middlware Tests", () => {
       const req = mocks.mockReq(mocks.mockLeadWebhook(), null, null, null, header)
       const next = mocks.mockNext()
       const { error } = invalidParamError('x-hub-signature')
-      await webhookMiddleware.checkSignature(req, res, next)
+      await xHubSignatureMiddleware.checkSignature(req, res, next)
       expect(res.status).toHaveBeenCalledWith(400)
       expect(res.json).toHaveBeenCalledWith(error)
     })
@@ -53,7 +53,7 @@ describe("WEBHOOK Middlware Tests", () => {
       }
       const req = mocks.mockReq(body, null, null, null, header)
       const next = mocks.mockNext()
-      await webhookMiddleware.checkSignature(req, res, next)
+      await xHubSignatureMiddleware.checkSignature(req, res, next)
       expect(res.status).not.toHaveBeenCalledWith(400)
       expect(next).toHaveBeenCalled()
     })
