+++ src/controllers/webhookController/index.js
@@ -1,23 +1,23 @@
 const { Router } = require('express')
 const { internalError } = require('../../helpers/Protocols')
 const { serverError } = require('../../helpers/Errors')
-const { LeadWebhookService } = require('../../services')
+const { WebhookService } = require('../../services')
 const ServiceException = require('../../helpers/Exceptions/ServiceException')
-const { WebhookMiddleware } = require('../../middlewares')
+const { XHubSignatureMiddleware } = require('../../middlewares')
 
 class LeadWebhookController {
   routes = Router()
   basePath = "/leads/facebook"
 
   constructor() {
-    this._webhookMiddleware = new WebhookMiddleware()
+    this._xHubSignatureMiddleware = new XHubSignatureMiddleware()
     this._load()
   }
 
   async _load() {
     this.routes.get(`${this.basePath}/sub`, this.subscrive)
     this.routes.route(`${this.basePath}`)
-      .all(this._webhookMiddleware.checkSignature)
+      .all(this._xHubSignatureMiddleware.checkSignature)
       .post(this.addLead)
 
   }
@@ -25,8 +25,8 @@ class LeadWebhookController {
 
   async subscrive(request, response) {
     try {
-      const leadWebhookService = new LeadWebhookService()
-      const result = await leadWebhookService.subscrive(request.query)
+      const webhookService = new WebhookService()
+      const result = await webhookService.subscrive(request.query)
       return response.status(200).json(result)
     } catch (err) {
       if (err instanceof ServiceException) {
@@ -43,8 +43,8 @@ class LeadWebhookController {
 
   async addLead(request, response){
     try {
-      const leadWebhookService = new LeadWebhookService()
-      const result = await leadWebhookService.addLead(request.body)
+      const webhookService = new WebhookService()
+      const result = await webhookService.addLead(request.body)
       return response.status(200).json(result)
     } catch (err) {
       if (err instanceof ServiceException) {
