+++ src/services/index.js
@@ -6,7 +6,7 @@ module.exports = {
   TaskService: require('./TaskService'),
   UserCampaignService: require('./UserCampaignService'),
   UserService: require('./UserService'),
-  LeadWebhookService: require('./LeadWebhookService'),
   WebhookService: require('./WebhookService'),
+  XHubSignatureService: require('./XHubSignatureService'),
   TokenService: require('./TokenService')
 }
